Clean up server.js (#421)
Server.js has become larger over time taking on a bunch of functions that would be better split out into utilities.  This streamlines server.js back to basics making it easier to see what's going on.
